Prevent calling a stabbed method within a transaction
With a new setting `within_transaction: false` you can prevent
calling a stabbed method within a database transaction.

```yaml
---
- class: Foo
  chain:
    - bar
    - baz
  within_transaction: false
  actions:
    - return: 0
```

If you call the method inside a transaction, the call will raise
an exception `::Isolator::UnsafeOperationError` pointing to
the method `Foo.bar.baz(*)`.

Notice that this setting is argument-specific. You can allow it
to be called inside a transaction with some arguments, and
<explicitly> forbid for some other arguments at the same time.