Refactor the gem to add method `call_fixture`
In this commit the gem was massively re-written to unify
changes including both seeds and stubs. Except for the code clearance,
this refactoring was needed to add the method `call_fixture`
which can load stubs and seeds from the same fixture.

Now you don't need to separate database preparation and
environment preparation between different files, but define
them in the single one like:

```yaml
---
- type: user
  params:
    id: 1
    name: Andrew

- class: GetUser
  chain: call
  attributes:
    - :id: 1
  actions:
    - return:
        id: 1
        name: Andrew
        admin: true
```

This fixture can be applied by method `call_fixture`, and the legacy
methods `stub_fixture` and `seed_fixture` are kept as its aliases
for the backward compatibility.